+++ packages/ak-dropdown/CHANGELOG.md
@@ -1,5 +1,25 @@
-<a name="0.0.0"></a>
-# 0.0.0 (2016-08-25)
+<a name="0.4.0"></a>
+# 0.4.0 (2016-08-29)
+
+
+
+<a name="ak-dropdown@0.3.0"></a>
+# ak-dropdown@0.3.0 (2016-08-29)
+
+
+### Features
+
+* Make babel-runtime part of the package, so it can be de-duplicated on build by the consumer ([9863540](https://bitbucket.org/atlassian/atlaskit/commits/9863540))
+
+
+
+<a name="ak-dropdown@0.2.0"></a>
+# ak-dropdown@0.2.0 (2016-08-26)
+
+
+
+<a name="ak-dropdown@0.1.0"></a>
+# ak-dropdown@0.1.0 (2016-08-26)
 
 
 ### Bug Fixes
@@ -9,15 +29,18 @@
 
 ### Features
 
-* add element to the keypresshandler ([a772d2b](https://bitbucket.org/atlassian/atlaskit/commits/a772d2b))
 * **bump:** BREAKING CHANGE: just to make it release v1 ([b2e2575](https://bitbucket.org/atlassian/atlaskit/commits/b2e2575))
 * **component:** add select item functionality ([15d175e](https://bitbucket.org/atlassian/atlaskit/commits/15d175e))
 * **component:** ak-dropdown-item finished ([720ff0e](https://bitbucket.org/atlassian/atlaskit/commits/720ff0e))
 * **component:** basic keyboard support ([64ca37a](https://bitbucket.org/atlassian/atlaskit/commits/64ca37a))
+* **component:** change selected color ([c7c83d1](https://bitbucket.org/atlassian/atlaskit/commits/c7c83d1))
 * **component:** close dropdown on select item ([c58e612](https://bitbucket.org/atlassian/atlaskit/commits/c58e612))
 * **component:** initial commit for ak-dropdown ([34a2d22](https://bitbucket.org/atlassian/atlaskit/commits/34a2d22))
+* **component:** item now have link with href ([ea4f33f](https://bitbucket.org/atlassian/atlaskit/commits/ea4f33f))
 * **component:** make dropdown close on click outside and esc button ([4e4510f](https://bitbucket.org/atlassian/atlaskit/commits/4e4510f))
 * **component:** open property for the dropdown ([84ba2fa](https://bitbucket.org/atlassian/atlaskit/commits/84ba2fa))
+* add element to the keypresshandler ([a772d2b](https://bitbucket.org/atlassian/atlaskit/commits/a772d2b))
+* Fix various bugs in ak-editor-bitbucket, upgrade skatejs everywhere. ([122b23a](https://bitbucket.org/atlassian/atlaskit/commits/122b23a))
 * **component:** trigger now supports keyboard interactions ([fa30154](https://bitbucket.org/atlassian/atlaskit/commits/fa30154))