+++ packages/ak-editor-toolbar-lists/CHANGELOG.md
@@ -1,5 +1,27 @@
-<a name="0.0.2"></a>
-## 0.0.2 (2016-08-25)
+<a name="1.2.0"></a>
+# 1.2.0 (2016-08-29)
+
+
+
+<a name="ak-editor-toolbar-lists@0.2.0"></a>
+# ak-editor-toolbar-lists@0.2.0 (2016-08-29)
+
+
+### Features
+
+* Make babel-runtime part of the package, so it can be de-duplicated on build by the consumer ([9863540](https://bitbucket.org/atlassian/atlaskit/commits/9863540))
+* Update 0.x editor components to 1.x ([3324adb](https://bitbucket.org/atlassian/atlaskit/commits/3324adb))
+
+
+
+<a name="ak-editor-toolbar-lists@0.1.0"></a>
+# ak-editor-toolbar-lists@0.1.0 (2016-08-26)
+
+
+### Features
+
+* Fix various bugs in ak-editor-bitbucket, upgrade skatejs everywhere. ([122b23a](https://bitbucket.org/atlassian/atlaskit/commits/122b23a))
+* Initial integration of lists plugin. ([fe33246](https://bitbucket.org/atlassian/atlaskit/commits/fe33246))