+++ packages/ak-editor-toolbar-text-formatting/CHANGELOG.md
@@ -1,10 +1,35 @@
-<a name="2.0.2"></a>
-## 2.0.2 (2016-08-25)
+<a name="2.3.0"></a>
+# 2.3.0 (2016-08-29)
+
+
+
+<a name="ak-editor-toolbar-text-formatting@2.2.0"></a>
+# ak-editor-toolbar-text-formatting@2.2.0 (2016-08-29)
+
+
+### Features
+
+* Make babel-runtime part of the package, so it can be de-duplicated on build by the consumer ([9863540](https://bitbucket.org/atlassian/atlaskit/commits/9863540))
+
+
+
+<a name="ak-editor-toolbar-text-formatting@2.1.0"></a>
+# ak-editor-toolbar-text-formatting@2.1.0 (2016-08-26)
+
+
+### Features
+
+* Initial integration of lists plugin. ([fe33246](https://bitbucket.org/atlassian/atlaskit/commits/fe33246))
 
 
 
 <a name="ak-editor-toolbar-text-formatting@2.0.2"></a>
 ## ak-editor-toolbar-text-formatting@2.0.2 (2016-08-25)
+
+
+### Features
+
+* Fix various bugs in ak-editor-bitbucket, upgrade skatejs everywhere. ([122b23a](https://bitbucket.org/atlassian/atlaskit/commits/122b23a))